+++ crates/weaver_forge/src/lib.rs
@@ -842,6 +842,7 @@ mod tests {
     use weaver_diff::diff_dir;
     use weaver_resolver::{LoadedSemconvRegistry, SchemaResolver};
     use weaver_semconv::registry_repo::RegistryRepo;
+    use weaver_semconv::schema_url::SchemaUrl;
 
     use crate::config::{ApplicationMode, CaseConvention, Params, TemplateConfig, WeaverConfig};
     use crate::debug::print_dedup_errors;
@@ -855,12 +856,12 @@ mod tests {
         cli_params: Params,
         ignore_non_fatal_errors: bool,
     ) -> (TemplateEngine, ResolvedRegistry, PathBuf, PathBuf) {
-        let registry_id = "default";
+        let schema_url = Some(SchemaUrl::try_new("https://default/1.0.0".to_owned()).unwrap());
         let path: VirtualDirectoryPath = "data/registry"
             .try_into()
             .expect("Invalid virtual directory path string");
         let repo =
-            RegistryRepo::try_new(registry_id, &path).expect("Failed to construct repository");
+            RegistryRepo::try_new(schema_url, &path).expect("Failed to construct repository");
         let registry_result = SchemaResolver::load_semconv_repository(repo, false);
         // SemConvRegistry::try_from_path_pattern(registry_id, "data/*.yaml");
         let registry = if ignore_non_fatal_errors {
@@ -1056,12 +1057,12 @@ mod tests {
         });
         engine.target_config.templates = Some(templates);
 
-        let registry_id = "default";
         let path: VirtualDirectoryPath = "data/registry"
             .try_into()
             .expect("Invalid virtual directory path string");
+        let schema_url = Some(SchemaUrl::try_new("https://default/1.0.0".to_owned()).unwrap());
         let repo =
-            RegistryRepo::try_new(registry_id, &path).expect("Failed to construct repository");
+            RegistryRepo::try_new(schema_url, &path).expect("Failed to construct repository");
         let loaded = SchemaResolver::load_semconv_repository(repo, false)
             .into_result_with_non_fatal()
             .expect("Failed to load registry")
@@ -1186,12 +1187,12 @@ mod tests {
 
     #[test]
     fn test_comment_format() {
-        let registry_id = "default";
         let path: VirtualDirectoryPath = "data/mini_registry_for_comments"
             .try_into()
             .expect("Invalid virtual directory path string");
+        let schema_url = Some(SchemaUrl::try_new("https://default/1.0.0".to_owned()).unwrap());
         let repo =
-            RegistryRepo::try_new(registry_id, &path).expect("Failed to construct repository");
+            RegistryRepo::try_new(schema_url, &path).expect("Failed to construct repository");
         let loaded = SchemaResolver::load_semconv_repository(repo, false)
             .into_result_with_non_fatal()
             .expect("Failed to load registry")