+++ crates/weaver_resolver/src/registry.rs
@@ -851,6 +851,7 @@ mod tests {
     use std::error::Error;
     use std::fs::OpenOptions;
     use std::path::PathBuf;
+    use weaver_semconv::schema_url::SchemaUrl;
 
     use glob::glob;
     use serde::Serialize;
@@ -918,12 +919,12 @@ mod tests {
             let observed_output_dir = PathBuf::from(format!("observed_output/{test_dir}"));
             std::fs::create_dir_all(observed_output_dir.clone())
                 .expect("Failed to create observed output directory");
-            let registry_id = "default";
+            let schema_url = Some(SchemaUrl::try_new("https://default/0.1.0".to_owned()).unwrap());
             let location: VirtualDirectoryPath = format!("{test_dir}/registry")
                 .try_into()
                 .expect("Failed to parse file directory");
             let loaded = SchemaResolver::load_semconv_repository(
-                RegistryRepo::try_new(registry_id, &location).expect("Failed to load registry"),
+                RegistryRepo::try_new(schema_url, &location).expect("Failed to load registry"),
                 true,
             )
             .ignore(|e| {
@@ -1104,8 +1105,6 @@ groups:
 
     #[test]
     fn test_api_usage() -> Result<(), Box<dyn Error>> {
-        let registry_id = "local";
-
         // Load a semantic convention registry from a local directory.
         // Note: A method is also available to load a registry from a git
         // repository.
@@ -1113,7 +1112,10 @@ groups:
         let path = VirtualDirectoryPath::LocalFolder {
             path: "data/registry-test-7-spans/registry".to_owned(),
         };
-        let repo = RegistryRepo::try_new(registry_id, &path)?;
+
+        let schema_url =
+            Some(SchemaUrl::try_new("https://local/registry/1.0.0".to_owned()).unwrap());
+        let repo = RegistryRepo::try_new(schema_url, &path)?;
         let loaded =
             SchemaResolver::load_semconv_repository(repo, true).into_result_failing_non_fatal()?;
         let resolved_schema =