+++ crates/weaver_semconv/src/registry_repo.rs
@@ -6,13 +6,18 @@ use std::default::Default;
 use std::path::{Path, PathBuf};
 use std::sync::Arc;
 
-use crate::manifest::RegistryManifest;
+use crate::manifest::{Dependency, RegistryManifest};
+use crate::schema_url::SchemaUrl;
 use crate::Error;
 use weaver_common::vdir::{VirtualDirectory, VirtualDirectoryPath};
-use weaver_common::{get_path_type, log_info};
+use weaver_common::{get_path_type, log_info, log_warn};
+
+/// The name of the legacy registry manifest file.
+#[deprecated(note = "The registry manifest file is renamed to `manifest.yaml`.")]
+pub const LEGACY_REGISTRY_MANIFEST: &str = "registry_manifest.yaml";
 
 /// The name of the registry manifest file.
-pub const REGISTRY_MANIFEST: &str = "registry_manifest.yaml";
+pub const REGISTRY_MANIFEST: &str = "manifest.yaml";
 
 /// A semantic convention registry repository that can be:
 /// - A definition repository, which is one of:
@@ -21,10 +26,11 @@ pub const REGISTRY_MANIFEST: &str = "registry_manifest.yaml";
 ///   - Initialized from a Git archive
 /// - A published repository, which is a manifest file
 ///   that denotes where to find aspects of the registry.
-#[derive(Default, Debug, Clone)]
+#[derive(Debug, Clone)]
 pub struct RegistryRepo {
-    // A unique identifier for the registry (e.g. main, baseline, etc.)
-    id: Arc<str>,
+    /// The schema URL associated with the registry
+    /// May be derived from the manifest or the registry name and version if the manifest is not present.
+    schema_url: SchemaUrl,
 
     // A virtual directory containing the registry.
     registry: VirtualDirectory,
@@ -34,30 +40,65 @@ pub struct RegistryRepo {
 }
 
 impl RegistryRepo {
-    /// Creates a new `RegistryRepo` from a `RegistryPath` object that
-    /// specifies the location of the registry.
-    pub fn try_new(
-        registry_id_if_no_manifest: &str,
-        registry_path: &VirtualDirectoryPath,
-    ) -> Result<Self, Error> {
-        let mut registry_repo = Self {
-            id: Arc::from(registry_id_if_no_manifest),
-            registry: VirtualDirectory::try_new(registry_path)
-                .map_err(Error::VirtualDirectoryError)?,
-            manifest: None,
-        };
-        if let Some(manifest) = registry_repo.manifest_path() {
-            let registry_manifest = RegistryManifest::try_from_file(manifest)?;
-            registry_repo.id = Arc::from(registry_manifest.name.as_str());
-            registry_repo.manifest = Some(registry_manifest);
-        }
-        Ok(registry_repo)
+    /// Creates a new `RegistryRepo` from a `Dependency` object that specifies the schema URL and path.
+    pub fn try_new_dependency(dependency: &Dependency) -> Result<Self, Error> {
+        let path = dependency.registry_path.clone().unwrap_or_else(|| {
+            // If no registry path is provided, we assume it's the same schema_url.
+            VirtualDirectoryPath::RemoteArchive {
+                url: dependency.schema_url.to_string(),
+                sub_folder: None,
+            }
+        });
+        Self::try_new(Some(dependency.schema_url.clone()), &path)
     }
 
-    /// Returns the unique identifier for the registry.
+    /// Creates a new `RegistryRepo` from a schema URL and `RegistryPath` object that
+    /// specifies the location of the registry.
+    /// If there is no manifest and schema URL is not provided, registry
+    /// name and version are set to "unknown".
+    pub fn try_new(
+        schema_url: Option<SchemaUrl>,
+        registry_path: &VirtualDirectoryPath,
+    ) -> Result<Self, Error> {
+        let registry =
+            VirtualDirectory::try_new(registry_path).map_err(Error::VirtualDirectoryError)?;
+        // Try to load manifest
+        if let Some(manifest_path) = {
+            // We need a temporary RegistryRepo to call manifest_path
+            let temp_repo = Self {
+                schema_url: SchemaUrl::new_unknown(),
+                registry: registry.clone(),
+                manifest: None,
+            };
+            temp_repo.manifest_path()
+        } {
+            let registry_manifest = RegistryManifest::try_from_file(manifest_path)?;
+            Ok(Self {
+                schema_url: registry_manifest.schema_url.clone(),
+                registry,
+                manifest: Some(registry_manifest),
+            })
+        } else {
+            // No manifest
+            let schema_url_combined = schema_url.unwrap_or_else(SchemaUrl::new_unknown);
+            Ok(Self {
+                schema_url: schema_url_combined.clone(),
+                registry,
+                manifest: None,
+            })
+        }
+    }
+
+    /// Returns the registry name (from manifest if present, otherwise top-level field).
     #[must_use]
-    pub fn id(&self) -> Arc<str> {
-        self.id.clone()
+    pub fn name(&self) -> Arc<str> {
+        self.schema_url.name().into()
+    }
+
+    /// Returns the registry version (from manifest if present, otherwise top-level field).
+    #[must_use]
+    pub fn version(&self) -> Arc<str> {
+        self.schema_url.version().into()
     }
 
     /// Returns the local path to the semconv registry.
@@ -78,27 +119,27 @@ impl RegistryRepo {
         self.manifest.as_ref()
     }
 
-    /// Returns the resolved schema URL, if available in the manifest.
+    /// Returns the resolved schema URI, if available in the manifest.
     #[must_use]
-    pub fn resolved_schema_url(&self) -> Option<VirtualDirectoryPath> {
+    pub fn resolved_schema_uri(&self) -> Option<VirtualDirectoryPath> {
         let manifest = self.manifest.as_ref()?;
-        let resolved_url: &str = manifest.resolved_schema_url.as_ref()?;
-        match get_path_type(resolved_url) {
+        let resolved_uri: &str = manifest.resolved_schema_uri.as_ref()?;
+        match get_path_type(resolved_uri) {
             weaver_common::PathType::RelativePath => {
-                // We need to understand if the manifest URL is the same as the registry URL.
+                // We need to understand if the manifest URI is the same as the registry URI.
                 let vdir_was_manifest_file = self.manifest_path()? == self.registry.path();
                 Some(self.registry.vdir_path().map_sub_folder(|path| {
                     if vdir_was_manifest_file {
                         match Path::new(&path).parent() {
-                            Some(parent) => format!("{}/{resolved_url}", parent.display()),
+                            Some(parent) => format!("{}/{resolved_uri}", parent.display()),
                             None => "".to_owned(),
                         }
                     } else {
-                        format!("{path}/{resolved_url}")
+                        format!("{path}/{resolved_uri}")
                     }
                 }))
             }
-            _ => resolved_url.try_into().ok(),
+            _ => resolved_uri.try_into().ok(),
         }
     }
 
@@ -111,18 +152,42 @@ impl RegistryRepo {
             return Some(self.registry.path().to_path_buf());
         }
         let manifest_path = self.registry.path().join(REGISTRY_MANIFEST);
+        let legacy_path = self.registry.path().join(LEGACY_REGISTRY_MANIFEST);
         if manifest_path.exists() {
             log_info(format!(
                 "Found registry manifest: {}",
                 manifest_path.display()
             ));
             Some(manifest_path)
+        } else if legacy_path.exists() {
+            log_warn(format!(
+                "Found registry manifest: {}. Please rename file to {}, as the old name is deprecated and won't be supported in future versions.",
+                legacy_path.display(),
+                REGISTRY_MANIFEST
+            ));
+            Some(legacy_path)
         } else {
             log_info(format!(
                 "No registry manifest found: {}",
                 manifest_path.display()
             ));
             None
+        }
+    }
+
+    /// Returns the registry schema URL.
+    #[must_use]
+    pub fn schema_url(&self) -> SchemaUrl {
+        self.schema_url.clone()
+    }
+}
+
+impl Default for RegistryRepo {
+    fn default() -> Self {
+        Self {
+            schema_url: SchemaUrl::new_unknown(),
+            registry: VirtualDirectory::default(),
+            manifest: None,
         }
     }
 }
@@ -147,7 +212,7 @@ mod tests {
         let registry_path = VirtualDirectoryPath::LocalFolder {
             path: "../../crates/weaver_codegen_test/semconv_registry".to_owned(),
         };
-        let repo = RegistryRepo::try_new("main", &registry_path).unwrap();
+        let repo = RegistryRepo::try_new(None, &registry_path).unwrap();
         let repo_path = repo.path().to_path_buf();
         assert!(repo_path.exists());
         assert!(
@@ -167,13 +232,13 @@ mod tests {
             path: "tests/published_repository/resolved/1.0.0".to_owned(),
         };
         let repo =
-            RegistryRepo::try_new("main", &registry_path).expect("Failed to load test repository.");
+            RegistryRepo::try_new(None, &registry_path).expect("Failed to load test repository.");
         let Some(manifest) = repo.manifest() else {
             panic!("Did not resolve manifest for repo: {repo:?}");
         };
-        assert_eq!(manifest.name, "resolved");
+        assert_eq!(manifest.name(), "resolved");
 
-        let Some(resolved_path) = repo.resolved_schema_url() else {
+        let Some(resolved_path) = repo.resolved_schema_uri() else {
             panic!(
                 "Should find a resolved schema path from manifest in {}",
                 repo.registry_path_repr()
@@ -189,8 +254,8 @@ mod tests {
             path: "tests/published_repository/resolved/2.0.0".to_owned(),
         };
         let repo =
-            RegistryRepo::try_new("main", &registry_path).expect("Failed to load test repository.");
-        let Some(resolved_path) = repo.resolved_schema_url() else {
+            RegistryRepo::try_new(None, &registry_path).expect("Failed to load test repository.");
+        let Some(resolved_path) = repo.resolved_schema_uri() else {
             panic!(
                 "Should find a resolved schema path from manifest in {}",
                 repo.registry_path_repr()
@@ -203,8 +268,8 @@ mod tests {
             path: "tests/published_repository/3.0.0".to_owned(),
         };
         let repo =
-            RegistryRepo::try_new("main", &registry_path).expect("Failed to load test repository.");
-        let Some(resolved_path) = repo.resolved_schema_url() else {
+            RegistryRepo::try_new(None, &registry_path).expect("Failed to load test repository.");
+        let Some(resolved_path) = repo.resolved_schema_uri() else {
             panic!(
                 "Should find a resolved schema path from manifest in {}",
                 repo.registry_path_repr()