+++ crates/weaver_semconv_gen/src/v2.rs
@@ -413,6 +413,7 @@ mod tests {
     use weaver_semconv::{
         attribute::{AttributeType, PrimitiveOrArrayTypeSpec},
         group::InstrumentSpec,
+        schema_url::SchemaUrl,
         v2::{span::SpanName, CommonFields},
     };
 
@@ -451,8 +452,7 @@ mod tests {
     fn test_registry() -> ResolvedTelemetrySchema {
         ResolvedTelemetrySchema {
             file_format: "resolved/2.0.0".to_owned(),
-            schema_url: "todo/1.0.0".to_owned(),
-            registry_id: "main".to_owned(),
+            schema_url: SchemaUrl::try_new("https://todo/1.0.0".to_owned()).unwrap(),
             attribute_catalog: vec![Attribute {
                 key: "attr1".to_owned(),
                 r#type: AttributeType::PrimitiveOrArray(PrimitiveOrArrayTypeSpec::String),
@@ -466,7 +466,6 @@ mod tests {
                     attributes: vec![AttributeRef(0)],
                     common: CommonFields::default(),
                 }],
-                registry_url: "todo".to_owned(),
                 spans: vec![Span {
                     r#type: "trace.test".to_owned().into(),
                     kind: weaver_semconv::group::SpanKindSpec::Client,
@@ -570,7 +569,6 @@ mod tests {
                     },
                 }],
             },
-            registry_manifest: None,
         }
     }
 }